Wire up tx_valid, tx_invalid, and sighash test vectors — all passing
New test files:
- tx_tests.rs: 119 tx_valid + 93 tx_invalid vectors exercising the full
  pipeline (tx deserialization, UTXO lookup, VerifyScript, TxSignatureChecker)
- sighash_tests.rs: 500 legacy sighash computation tests

Bugs fixed:
- check_sequence version comparison: use unsigned (u32) cast, matching
  Bitcoin Core's uint32_t version field (version=-1 is 0xFFFFFFFF >= 2)
- OP_SUCCESSx scanner in witness.rs: rewrote to properly advance past
  push data without duplicated/broken logic
- Sighash test: strip OP_CODESEPARATOR before hashing (bitcoin crate's
  legacy_signature_hash doesn't do this, but Bitcoin Core does)
- BADTX handling: check structural validity (empty vin/vout, negative
  values, MAX_MONEY, duplicate inputs, null prevout, coinbase scriptSig
  size) since the bitcoin crate's decoder accepts some invalid txs

Test totals: 57 tests, 0 failures
Vector coverage: 1217 script + 500 sighash + 119 tx_valid + 93 tx_invalid
                 + 13 tapscript + 9 BIP341 inputs = 1951 verifications

+++ consensus/src/sighash.rs
@@ -221,8 +221,9 @@ impl SignatureChecker for TxSignatureChecker<'_> {
     fn check_sequence(&self, sequence: &ScriptNum) -> bool {
         let tx_sequence = self.tx.input[self.input_index].sequence.0;
 
-        // Tx version must be >= 2 for BIP68
-        if self.tx.version.0 < 2 {
+        // Tx version must be >= 2 for BIP68.
+        // Bitcoin Core uses uint32_t for version, so the comparison is unsigned.
+        if (self.tx.version.0 as u32) < 2 {
             return false;
         }
